refactor: move controllers folder and err folder to presentation folder
