+++ src/clients/client-catalog.js
@@ -1,4 +1,4 @@
-import { RequestError, ServerError } from '../app/err/errors.js'
+import { RequestError, ServerError } from '../presentation/err/errors.js'
 import * as HttpUtil from '../app/utils/request.js'
